# swagger.yaml
openapi: 3.0.0
info:
  title: Ecommerce-API
  contact: {}
  version: "1.0"
servers:
  - url: http://localhost:5000/api/v1
    variables: {}
paths:
...
  /categories/{categoryId}/subCategories:
    get:
      tags:
        - SUB CATEGORY
      summary: GET ALL SUBCATEGORIES
      operationId: GETALLSUBCATEGORIES
      parameters:
        - name: categoryId
          in: path
          description: ID of the category to retrieve its subcategories
          required: true
          schema:
            type: string
          example: 659c13de800a4f99852c9c14
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    post:
      tags:
        - SUB CATEGORY
      summary: CREATE SUBCATEGROY
      operationId: CREATESUBCATEGROY
      parameters:
        - name: categoryId
          in: path
          description: ID of the category that the subcategories will belong to
          required: true
          schema:
            type: string
          example: 659c13de800a4f99852c9c14
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CREATESUBCATEGROYREQUEST"
                - example:
                    name: white cats
            example:
              name: white cats
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /subCategories/{subCategoryId}:
    get:
      tags:
        - SUB CATEGORY
      summary: GET SINGLE SUBCATEGORY
      operationId: GETSINGLESUBCATEGORY
      parameters:
        - name: subCategoryId
          in: path
          description: ID of the subcategory to retrieve
          required: true
          schema:
            type: string
          example: 659c13de800a4f99852c9c14
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    patch:
      tags:
        - SUB CATEGORY
      summary: UPDATE SUBCATEGORY
      operationId: UPDATESUBCATEGORY
      parameters:
        - name: subCategoryId
          in: path
          description: ID of the subcategory to update
          required: true
          schema:
            type: string
          example: 659c13de800a4f99852c9c14
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UPDATESUBCATEGORYREQUEST"
                - example:
                    name: DJ Cats
            example:
              name: DJ Cats
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    delete:
      tags:
        - SUB CATEGORY
      summary: DELETE SUBCATEGORY
      operationId: DELETESUBCATEGORY
      parameters:
        - name: subCategoryId
          in: path
          description: ID of the subcategory to delete
          required: true
          schema:
            type: string
          example: 659c13de800a4f99852c9c14
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    UPDATESUBCATEGORYREQUEST:
      title: UPDATESUBCATEGORYREQUEST
      required:
        - name
      type: object
      properties:
        name:
          type: string
      example:
        name: DJ Cats
    CREATESUBCATEGROYREQUEST:
      title: CREATESUBCATEGROYREQUEST
      required:
        - name
      type: object
      properties:
        name:
          type: string
      example:
        name: white cats
...